video-transcript/requirements_pyannote.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# 基础
numpy==1.26.4
soundfile==0.12.1
ffmpeg-python==0.2.0
huggingface_hub==0.23.2

#PyTorchCPU
torch==2.2.2
torchaudio==2.2.2

# pyannote适配3.1
pyannote.audio==3.1.1